Source Code/requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.6.0
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
fastcore==1.5.29
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.6
importlib-metadata==6.11.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
namex==0.0.8
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.23.5
opencv-contrib-python==4.9.0.80
opencv-contrib-python-headless==4.8.1.78
opencv-python==4.10.0.82
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==1.3.4
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
preshed==3.0.9
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
reportlab==4.0.4
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.25.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.26.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
validators==0.21.2
wasabi==1.1.2
watchdog==4.0.0
wavio==0.0.9
weasel==0.3.4
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==1.7.2
zipp==3.18.1